Updates (#142)
* 2024-01-25 updates

* 2024-07-11

* 2024-07-16

* 2025-02-03

* 2025-02-10

* 2025-04-08

* Removed opportunities, no more opportunites for anyone! MWHAHAHAHA

* version update of package

+++ EVEStandard/Models/CharacterContactNotification.cs
@@ -19,7 +19,7 @@ namespace EVEStandard.Models
         /// </summary>
         /// <value>notification_id integer</value>
         [JsonPropertyName("notification_id")]
-        public int NotificationId { get; set; }
+        public long NotificationId { get; set; }
 
         /// <summary>
         /// send_date string
